pyproject.toml
[project]
name = "synapse-deepagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "apscheduler>=3.11.1",
    "asyncpg>=0.31.0",
    "bcrypt==4.3.0",
    "bedrock-agentcore>=1.0.7",
    "bedrock-agentcore-starter-toolkit>=0.2.0",
    "boto3>=1.41.5",
    "bs4>=0.0.2",
    "deepagents>=0.2.8",
    "docx>=0.2.4",
    "dotenv>=0.9.9",
    "faiss-cpu>=1.13.1",
    "fastapi>=0.122.0",
    "fpdf>=1.7.2",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.43.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.2",
    "hf-xet>=1.2.0",
    "ipython>=9.7.0",
    "langchain>=1.1.0",
    "langchain-astradb>=1.0.0",
    "langchain-aws>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-groq>=1.1.0",
    "langchain-huggingface>=1.1.0",
    "langchain-mcp-adapters>=0.1.14",
    "langchain-mongodb>=0.9.0",
    "langchain-openai>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.4",
    "langgraph-api>=0.5.27",
    "langgraph-checkpoint-mongodb>=0.3.0",
    "langgraph-cli[inmem]>=0.4.7",
    "langgraph-store-mongodb>=0.1.1",
    "matplotlib>=3.10.7",
    "nest-asyncio>=1.6.0",
    "numpy>=2.3.5",
    "odfpy>=1.4.1",
    "pandas>=2.3.3",
    "passlib[bcrypt]==1.7.4",
    "pillow>=12.0.0",
    "prophet>=1.2.1",
    "psutil>=7.1.3",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.10.1",
    "pymongo>=4.15.4",
    "pypdf>=6.4.0",
    "pypdf2>=3.0.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "pytz>=2025.2",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "resend>=2.19.0",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.44",
    "sqlparse>=0.5.3",
    "statsmodels>=0.14.5",
    "streamlit>=1.51.0",
    "tiktoken>=0.12.0",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]
